rename adventure.py to app.py
