+++ DungeonAPI/__init__.py
@@ -1,5 +1,5 @@
 """Entry point for our DungeonAPI flask app"""
 
-from .adventure import create_app
+from .app import create_app
 
 APP = create_app()
